# openapi.yaml
openapi: 3.0.3
info:
  title: Tennis Court Finder API
  description: |
    Monitors tennis court availability across integrated booking systems and sends
    real-time alerts to users when requested court times become available.

    ## Authentication Flow
    1. User requests a one-time password (OTP) sent to their email
    2. User verifies the OTP
    3. Server returns a JWT stored in an HTTP-only cookie
    4. All authenticated endpoints read identity from the JWT cookie
  version: 0.1.0
  contact:
    name: Tennis Court Finder
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List the authenticated user's notification subscriptions
      security:
        - cookieAuth: []
      parameters:
        - name: active
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by active/inactive status
        - name: club_id
          in: query
          required: false
          schema:
            type: string
            example: seb-arena
          description: Filter by club slug
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of notification subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscriptionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a new notification subscription
      description: |
        Subscribe to be notified when a time slot matching your criteria becomes
        available or changes status.

        **One-time**: provide `specific_dates` to watch for slots on particular dates.
        **Recurring**: set `is_recurring` to true and provide `days_of_week` to be
        notified on a weekly schedule (e.g. every Tuesday and Thursday).

        You can combine club, court, surface, and time filters to narrow the match.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSubscriptionCreate"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/notifications/{notification_id}:
    get:
      operationId: getNotification
      tags: [notifications]
      summary: Get a specific notification subscription
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/NotificationIdParam"
      responses:
        "200":
          description: Notification subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateNotification
      tags: [notifications]
      summary: Fully update a notification subscription
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/NotificationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSubscriptionUpdate"
      responses:
        "200":
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification subscription
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/NotificationIdParam"
      responses:
        "204":
          description: Subscription deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/notifications/{notification_id}/toggle:
    patch:
      operationId: toggleNotification
      tags: [notifications]
      summary: Activate or deactivate a notification subscription
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/NotificationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationToggle"
      responses:
        "200":
          description: Subscription toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSubscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/notifications/{notification_id}/logs:
    get:
      operationId: listNotificationLogs
      tags: [notifications]
      summary: List sent notification logs for a subscription
      description: |
        Returns the history of notifications that were actually sent for this
        subscription – useful for auditing and debugging.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/NotificationIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated notification logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationLogListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

# ═══════════════════════════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════════════════════════
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: JWT session token stored in an HTTP-only cookie

  # ── Reusable Parameters ──────────────────────────────────────────────
  parameters:
...
    NotificationIdParam:
      name: notification_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the notification subscription
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSizeParam:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page

  # ── Reusable Responses ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ── Schemas ─────────────────────────────────────────────────────────
  schemas:
...
    SurfaceType:
      type: string
      enum: [hard, clay, carpet, grass, artificial_grass]
      description: Court surface material
    CourtType:
      type: string
      enum: [indoor, outdoor]
      description: Whether the court is indoors or outdoors
    SlotStatus:
      type: string
      enum: [free, for_sale, booked]
      description: |
        Availability status of a time slot:
        - **free** – available for booking
        - **for_sale** – a previously booked slot released for resale
        - **booked** – currently reserved
    DayOfWeek:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]

    # ── Generic ─────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: Resource not found
    ValidationErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
    PaginationMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 3

    # ── Auth ────────────────────────────────────────────────────────
...
    TimeSlot:
      type: object
      required: [id, court_id, club_id, court_name, start_time, end_time, status]
      properties:
        id:
          type: string
          format: uuid
        court_id:
          type: string
          format: uuid
        club_id:
          type: string
          description: Slug of the parent club
          example: seb-arena
        court_name:
          type: string
          example: Court 1
          description: Denormalized court name for display convenience
        surface_type:
          $ref: "#/components/schemas/SurfaceType"
        court_type:
          $ref: "#/components/schemas/CourtType"
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          example: 60
        status:
          $ref: "#/components/schemas/SlotStatus"
        price:
          type: number
          format: float
          nullable: true
          example: 25.00
        currency:
          type: string
          nullable: true
          example: EUR
...
    NotificationSubscription:
      type: object
      required:
        - id
        - club_id
        - notify_on_statuses
        - is_recurring
        - active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        club_id:
          type: string
          description: Slug of the club being watched
          example: seb-arena
        club_name:
          type: string
          description: Denormalized club name for display
          example: City Tennis Club
        court_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
          description: Specific courts to watch (null or empty = all courts in the club)
        surface_types:
          type: array
          items:
            $ref: "#/components/schemas/SurfaceType"
          nullable: true
          description: Filter by surface type (null = any)
        court_types:
          type: array
          items:
            $ref: "#/components/schemas/CourtType"
          nullable: true
          description: Filter by indoor/outdoor (null = any)
        notify_on_statuses:
          type: array
          items:
            $ref: "#/components/schemas/SlotStatus"
          minItems: 1
          description: Statuses that trigger a notification (e.g. [free, for_sale])
        time_from:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
          example: "18:00"
          description: Earliest time of day to watch (HH:MM, 24h)
        time_to:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
          example: "21:00"
          description: Latest time of day to watch (HH:MM, 24h)
        is_recurring:
          type: boolean
          description: Whether this is a weekly recurring watch
        days_of_week:
          type: array
          items:
            $ref: "#/components/schemas/DayOfWeek"
          nullable: true
          description: Days of the week to watch (required when is_recurring is true)
        specific_dates:
          type: array
          items:
            type: string
            format: date
          nullable: true
          description: Specific dates to watch (used when is_recurring is false)
        date_range_start:
          type: string
          format: date
          nullable: true
          description: Start watching from this date (inclusive)
        date_range_end:
          type: string
          format: date
          nullable: true
          description: Stop watching after this date (inclusive)
        active:
          type: boolean
          description: Whether this subscription is currently active
        match_count:
          type: integer
          description: Number of times this subscription has matched
          example: 5
        last_notified_at:
          type: string
          format: date-time
          nullable: true
          description: When the last notification was sent
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NotificationSubscriptionCreate:
      type: object
      required: [club_id, notify_on_statuses, is_recurring]
      properties:
        club_id:
          type: string
          description: Slug of the club to watch
          example: seb-arena
        court_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
        surface_types:
          type: array
          items:
            $ref: "#/components/schemas/SurfaceType"
          nullable: true
        court_types:
          type: array
          items:
            $ref: "#/components/schemas/CourtType"
          nullable: true
        notify_on_statuses:
          type: array
          items:
            $ref: "#/components/schemas/SlotStatus"
          minItems: 1
        time_from:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
          example: "18:00"
        time_to:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
          example: "21:00"
        is_recurring:
          type: boolean
        days_of_week:
          type: array
          items:
            $ref: "#/components/schemas/DayOfWeek"
          nullable: true
        specific_dates:
          type: array
          items:
            type: string
            format: date
          nullable: true
        date_range_start:
          type: string
          format: date
          nullable: true
        date_range_end:
          type: string
          format: date
          nullable: true
    NotificationSubscriptionUpdate:
      type: object
      required: [club_id, notify_on_statuses, is_recurring]
      properties:
        club_id:
          type: string
          description: Slug of the club to watch
          example: seb-arena
        court_ids:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
        surface_types:
          type: array
          items:
            $ref: "#/components/schemas/SurfaceType"
          nullable: true
        court_types:
          type: array
          items:
            $ref: "#/components/schemas/CourtType"
          nullable: true
        notify_on_statuses:
          type: array
          items:
            $ref: "#/components/schemas/SlotStatus"
          minItems: 1
        time_from:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
          example: "18:00"
        time_to:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
          example: "21:00"
        is_recurring:
          type: boolean
        days_of_week:
          type: array
          items:
            $ref: "#/components/schemas/DayOfWeek"
          nullable: true
        specific_dates:
          type: array
          items:
            type: string
            format: date
          nullable: true
        date_range_start:
          type: string
          format: date
          nullable: true
        date_range_end:
          type: string
          format: date
          nullable: true
    NotificationToggle:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
          description: Set to true to activate, false to deactivate
    NotificationSubscriptionListResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NotificationSubscription"
        meta:
          $ref: "#/components/schemas/PaginationMeta"

    # ── Notification Log ────────────────────────────────────────────
    NotificationLog:
      type: object
      required: [id, subscription_id, sent_at, channel, time_slot]
      properties:
        id:
          type: string
          format: uuid
        subscription_id:
          type: string
          format: uuid
        sent_at:
          type: string
          format: date-time
        channel:
          type: string
          enum: [email]
          description: Notification channel used
        time_slot:
          $ref: "#/components/schemas/TimeSlot"
          description: The time slot that triggered the notification
        status:
          type: string
          enum: [sent, failed]
        error_message:
          type: string
          nullable: true
    NotificationLogListResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NotificationLog"
        meta:
          $ref: "#/components/schemas/PaginationMeta"